Add method `find_to'.
The method `find_to' finds a node in the trie and remembers and returns
an array of all nodes in the trie visited in order to find the target.
This is useful for generating the a website's navigation bar: the
structure of the site is stored in the trie, using URLs as keys.  Using
`find_to' all elements that should be highlighted in the navigation bar
can be determined for a given URL.

Considering a trie with the following keys:
    abc
    ade
    bbb

Calling `find_to' with the argument `abc' would return:
    ['a','ab','abc']